requirements.txt
flask==2.1.3
flask-smorest
python-dotenv
SQLAlchemy==1.4.44
flask-sqlalchemy
jsonpickle
requests
Flask-SocketIO==5.3.2
flask-jwt-extended
passlib
flask-migrate
gunicorn
psycopg2
